% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMLEmode.R
\name{AMLEmode}
\alias{AMLEmode}
\title{Approximating the mode of a multivariate empirical distribution}
\usage{
AMLEmode(ABCsam)
}
\arguments{
\item{ABCsam}{(m x k) matrix: ABC sample, where m is the ABC sample size and k is the
number of parameters.}
}
\value{
A list containing the 4 approximate modes.
}
\description{
This function approximates the mode of a multivariate empirical distribution by means of:
\enumerate{
\item the sample mean
\item the product of the maxima of the univariate kernel densities estimated using the marginals
\item the maximum of the multivariate kernel density
\item the maximum of the product of the univariate kernel densities
}
}
\details{
The bandwidth is estimated via smoothed cross-validation
}
\examples{
mixFit <- DynMixfit(TN2016,0.005,5000,100,AMLE=TRUE)
res <- AMLEmode(mixFit$ABCsam)
}
\keyword{approximate}
\keyword{dynamic}
\keyword{likelihood.}
\keyword{maximum}
\keyword{mixture;}
