% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootsamples.R
\name{Bootsamples}
\alias{Bootsamples}
\title{Creating bootstrap samples from observed data and computing MLEs}
\usage{
Bootsamples(y, nboot)
}
\arguments{
\item{y}{numerical vector: observed random sample from the mixture.}

\item{nboot}{number of bootstrap replications}
}
\value{
A list with the following elements:

MLE: maximum likelihood estimates obtained from each bootstrap sample.

errors: number of times the MLE algorithm breaks down.
}
\description{
This function creates bootstrap samples of input data and fits a dynamic mixture via
standard maximum likelihood.
}
\details{
MLEs are computed by means of the optim function. When it breaks
down, the sample is discarded and a new one is generated. The function keeps
track of the number of times this happens.
}
\examples{
bootMLEs <- Bootsamples(TN2016,100)
}
\keyword{MLE;}
\keyword{bootstrap.}
\keyword{dynamic}
\keyword{mixture;}
\keyword{non-parametric}
