% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm_stat_M.R
\name{cvm_stat_M}
\alias{cvm_stat_M}
\title{Computing the Cramér - von Mises distance between two samples}
\usage{
cvm_stat_M(vec1, vec2, power)
}
\arguments{
\item{vec1}{(n x 1) vector: first sample.}

\item{vec2}{(n x 1) vector: second sample.}

\item{power}{power of the integrand. Equal to 2 when using the L2 distance}
}
\value{
out scalar: Cramér - von Mises distance between the two samples
}
\description{
This function Computing the Cramér - von Mises distance between two samples.
}
\details{
This function computes the Cramér - von Mises distance
between two samples. See Drovandi and Frazier (2022, p. 7).
}
\examples{
out = cvm_stat_M(runif(100),rnorm(100),2)
}
\references{
{
\insertRef{dro22}{LNPar}
}
}
\keyword{dynamic}
\keyword{mixture;}
\keyword{simulation.}
