% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpareto.r
\name{dpareto}
\alias{dpareto}
\title{density of a Pareto r.v.}
\usage{
dpareto(x, xmin, alpha)
}
\arguments{
\item{x}{numerical vector (>=xmin): values where the density has to be evaluated.}

\item{xmin}{positive scalar: Pareto scale parameter.}

\item{alpha}{positive scalar: Pareto shape parameter.}
}
\value{
Density of the Pareto distribution evaluated at x.
}
\description{
This function evaluates the density of a Pareto r.v.s
}
\examples{
parDens <- dpareto(5,4,2)
}
\keyword{density.}
