% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynloglik.R
\name{dynloglik}
\alias{dynloglik}
\title{Log-likelihood of a Lognormal-GPD dynamic mixture}
\usage{
dynloglik(x, y)
}
\arguments{
\item{x}{(6 by 1) numerical vector: values of CA1, CA2, meanlog, sdlog, xi, beta.}

\item{y}{vector: points where the function is evaluated.}
}
\value{
log-likelihood of the lognormal-GPD mixture evaluated at y.
}
\description{
This function evaluates the log-likelihood of a Lognormal-GPD dynamic mixture.
}
\examples{
fit <- dynloglik(x,y)
}
\keyword{dynamic}
\keyword{mixture.}
