% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPfit.R
\name{LPfit}
\alias{LPfit}
\title{Estimating a lognormal-Pareto mixture by maximizing the profile log-likelihood}
\usage{
LPfit(y, minRank, takeOut, pimax, nboot)
}
\arguments{
\item{y}{numerical vector: random sample from the mixture.}

\item{minRank}{integer: minimum possible rank of the threshold.}

\item{takeOut}{integer: minimum number of observations above the threshold (see Details).}

\item{pimax}{integer: prior probability threshold for estimation of a pure lognormal (see Details).}

\item{nboot}{number of bootstrap replications used for estimating the standard errors. If omitted, no standard errors are computed.}
}
\value{
A list with the following elements:

xmin: estimated threshold.

prior: estimated mixing weight.

postProb: matrix of posterior probabilities.

alpha: estimated Pareto shape parameter.

mu: estimated expectation of the lognormal distribution on the lognormal scale.

sigma: estimated standard deviation of the lognormal distribution on the lognormal scale.

loglik: maximized log-likelihood.

nit: number of iterations.

npareto: estimated number of Pareto observations.

bootstd: bootstrap standard errors of the estimators.
}
\description{
This function fits a lognormal-Pareto mixture by maximizing the profile log-likelihood.
}
\details{
Estimation is implemented as in Bee (2022). The method is applied to each possible threshold th (minRank <= th <= n-takeOut).
If the estimated prior probability is larger than pimin, the prior probability is set equal to 1, and a pure lognormal is estimated via MLE.
As of standard errors, at each bootstrap replication the mixture is estimated with thresholds equal to ys(minRank), ys(minRank+1),..., ys(n),
where n is the sample size and ys is the sample sorted in ascending order. The latter procedure is implemented via parallel computing.
If the algorithm does not converge in 1000 iterations, a message is displayed.
}
\examples{
mixFit <- LPfit(TN2016,90,5,0.99,0)
}
\references{
{
\insertRef{bee22}{LNPar}
}
}
\keyword{likelihood.}
\keyword{mixture;}
\keyword{profile}
