% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Metro2019}
\alias{Metro2019}
\title{Population estimate of the US metropolitan areas}
\format{
A numerical vector with 415 rows and 1 column.
}
\source{
\url{https://www.census.gov/data/datasets/time-series/demo/popest/2010s-total-metro-and-micro-statistical-\%20areas.html/}
}
\usage{
Metro2019
}
\description{
A dataset cotaining the 2019 population estimate, divided by 10000, of the
415 US metropolitan areas computed by the US Census Bureau.
}
\keyword{datasets}
