% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEBoot.R
\name{MLEBoot}
\alias{MLEBoot}
\title{Bootstrap standard errors for the estimators of a lognormal-Pareto mixture}
\usage{
MLEBoot(x, y, minRank, takeOut, p0, alpha0, mu0, Psi0)
}
\arguments{
\item{x}{list: sequence of integers 1,...,K, where K is the mumber of datasets. Set x = 1 in case
of a single dataset.}

\item{y}{numerical vector: observed sample.}

\item{minRank}{positive integer: minimum possible rank of the threshold.}

\item{takeOut}{integer: minimum number of observations above the threshold (see Details).}

\item{p0}{(0<p0<1): starting value of the mixing weight.}

\item{alpha0}{non-negative scalar: starting value of the Pareto shape parameter.}

\item{mu0}{scalar: starting value of the log-expectation of the lognormal distribution on the log scale.}

\item{Psi0}{non-negative scalar: starting value of the log-variance of the lognormal distribution on the log scale.}
}
\value{
Estimated parameters obtained from a bootstrap sample.
}
\description{
This function draws a bootstrap sample and uses it to estimate the parameters of a lognormal-Pareto mixture distribution. Since this is typically called by LPfit, see the help of LPfit for examples.
}
\details{
At each bootstrap replication, the mixture is estimated with thresholds equal to ys(minRank), ys(minRank+1),..., ys(n-takeOut),
where n is the sample size and ys is the sample in ascending order.
If the estimated prior probability is larger than pimin, the prior probability is set equal to 1, and a pure lognormal is estimated via MLE.
The function is typically called by LPfit (see the example below).
}
\examples{
mixFit <- LPfit(TN2016,90,5,10)
}
\references{
Bee, M. (2022), “On discriminating between lognormal and Pareto tail: a mixture-based approach”,
Advances in Data Analysis and Classification, https://doi.org/10.1007/s11634-022-00497-4
}
\keyword{EM}
\keyword{algorithm.}
\keyword{likelihood;}
\keyword{mixture;}
\keyword{profile}
