% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpareto.r
\name{rpareto}
\alias{rpareto}
\title{Random number generation for a Pareto r.v.}
\usage{
rpareto(n, xmin, alpha)
}
\arguments{
\item{n}{positive integer: number of simulated random numbers.}

\item{xmin}{scalar: Pareto scale parameter.}

\item{alpha}{non-negative scalar: Pareto shape parameter.}
}
\value{
n iid random numbers from the Pareto distribution.
}
\description{
This function simulates random numbers for a Pareto r.v.
}
\examples{
ySim <- rpareto(5,4,1.5)
}
\keyword{number}
\keyword{random}
\keyword{simulation.}
