% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TN2016}
\alias{TN2016}
\title{Number of employees in year 2016 in all the firms of the Trento district in Northern Italy.}
\format{
A numerical vector with 183 rows and 1 column.
}
\source{
\url{https://dati.trentino.it/dataset/addetti-in-imprese-attive/}
}
\usage{
TN2016
}
\description{
A dataset containing the number of employees in year 2016 in all the firms of the Trento district.
}
\keyword{datasets}
