% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpareto.r
\name{dpareto}
\alias{dpareto}
\title{density of a Pareto r.v.}
\usage{
dpareto(x, xmin, alpha)
}
\arguments{
\item{x}{numerical vector: values where the density has to be evaluated.}

\item{xmin}{scalar: Pareto scale parameter.}

\item{alpha}{non-negative scalar: Pareto shape parameter.}
}
\value{
Density of the Pareto distribution evaluated at x.
}
\description{
This function evaluates the density of a Pareto r.v.s
}
\examples{
parDens <- dPareto(5,4,1.5)
}
\keyword{density.}
