% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMixfit.R
\name{DynMixfit}
\alias{DynMixfit}
\title{Estimating a dynamic mixture by AMLE and MLE}
\usage{
DynMixfit(yObs, epsilon, k, bootreps)
}
\arguments{
\item{yObs}{numerical vector: observed random sample from the mixture.}

\item{epsilon}{non-negative scalar: scale parameter of the Markov kernel.}

\item{k}{non-negative integer: number of samples generated in the AMLE
approach, such that k*epsilon = ABC sample size.}

\item{bootreps}{non-negative integer: number of bootstrap replications.}
}
\value{
A list with the following elements:

AMLEpars (6 x 1) vector: approximate maximum likelihood estimates computed via sample mean,
maxima of the marginal kernel densities, maximum of the multivariate kernel densities,
maximum of the product of the marginal kernel densities.

MLEpars (7 x 1) vector: maximum likelihood estimates and
maximized log-likelihood.

MLE (bootreps x 6) matrix: maximum likelihood estimates obtained in
each bootstrap replication.

ABCsam ((k x epsilon) x 6) matrix: ABC sample.
}
\description{
This function fits a dynamic mixture by Approximate Maximum Likelihood
and by standard maximum likelihood.
Currently only implemented for the lognormal - generalized Pareto case.
}
\details{
Starting values for mu and sigma are the lognormal MLEs computed
with the observations below the median. Initial values for xi and
tau are the GPD MLEs obtained with the observations above the median.
For the location and scale parameter of the Cauchy, we respectively use the first quartile and log(sd(x)/2).
In AMLE, for the lognormal and GPD parameters, the support of the uniform
prior is set equal to the 99\% confidence interval of the bootstrap
distribution after discarding the outliers.
For the Cauchy parameters, the support is given by the range of the
bootstrap distribution after discarding the outliers.
Be aware that computing times are large when k and/or bootreps are large.
}
\examples{
mixFit <- DynMixfit(TN2016,0.005,5000,100)
}
\keyword{approximate}
\keyword{dynamic}
\keyword{likelihood.}
\keyword{maximum}
\keyword{mixture;}
