% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPfit.R
\name{LPfit}
\alias{LPfit}
\title{Estimating a lognormal-Pareto mixture by maximizing the profile log-likelihood}
\usage{
LPfit(y, minRank, nbootMLE)
}
\arguments{
\item{y}{numerical vector: random sample from the mixture.}

\item{minRank}{integer: minimum possible rank of the threshold.}
}
\value{
A list with the following elements:

xmin: estimated threshold.

prior: estimated mixing weight.

postProb: matrix of posterior probabilities.

alpha: estimated Pareto shape parameter.

mu: estimated expectation of the lognormal distribution on the lognormal scale.

sigma: estimated standard deviation of the lognormal distribution on the lognormal scale.

loglik: maximied log-likelihood.

nit: number of iterations.

npareto: estimated number of Pareto observations.

bootstd: bootstrap standard errors of the parameter estimators.
}
\description{
This function fits a lognormal-Pareto mixture by maximizing the profile log-likelihood.
}
\details{
At each bootstrap replication, the mixture is estimated with thresholds equal to ys(n-nthresh), ys(n-nthresh+1),..., ys(n),
where n is the sample size and ys is the sample in ascending order.
}
\examples{
mixFit <- LPfit(y,90,0)
}
\keyword{likelihood.}
\keyword{mixture;}
\keyword{profile}
