% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPfitEM.R
\name{LPfitEM}
\alias{LPfitEM}
\title{Estimating a lognormal-Pareto mixture via the ECME algorithm}
\usage{
LPfitEM(y, eps, maxiter, qxmin0)
}
\arguments{
\item{y}{numerical vector: random sample from the mixture.}

\item{minRank}{integer: minimum possible rank of the threshold.}

\item{nboot}{number of bootstrap replications used for estimating the standard errors. If omitted, no standard errors are computed.}
}
\value{
A list with the following elements:

xmin: estimated threshold.

prior: estimated mixing weight.

postProb: matrix of posterior probabilities.

alpha: estimated Pareto shape parameter.

mu: estimated expectation of the lognormal distribution on the lognormal scale.

sigma: estimated standard deviation of the lognormal distribution on the lognormal scale.

loglik: maximized log-likelihood.

nit: number of iterations.

npareto: estimated number of Pareto observations.

bootstd: bootstrap standard errors of the estimators.
}
\description{
This function fits a lognormal-Pareto mixture by means of the ECME algorithm.
}
\details{
Estimation of a lognormal-Pareto mixture via the ECME algorithm.
}
\examples{
mixFit <- LPfitEM(TN2016,0)


}
\keyword{ECME}
\keyword{algorithm.}
\keyword{mixture;}
