% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPtest.R
\name{LPtest}
\alias{LPtest}
\title{Testing for a Pareto tail}
\usage{
LPtest(x, n, muNull, sigmaNull, obsTest, minRank)
}
\arguments{
\item{x}{socket cluster}

\item{n}{sample size}

\item{muNull}{lognormal expected value under the null hypothesis.}

\item{sigmaNull}{lognormal standard deviation under the null hypothesis.}

\item{minRank}{minimum possible rank of the threshold.}

\item{ObsTest:}{observed value of the test}
}
\value{
A list with the following elements:

LR: observed value of the llr test.
}
\description{
This function draws a bootstrap sample from the null (lognormal) distribution and computes the test for the null hypothesis of
a pure lognormal distribution versus the alternative of a lognormal-Pareto
mixture. To be only called from ParallelTest.
}
\examples{
minRank = 90
mixFit <- LPfit(TN2016,minRank,2)
ell1 <- mixFit$loglik
estNull <- c(mean(log(TN2016)),sd(log(TN2016)))
ellNull <- sum(log(dlnorm(TN2016,estNull[1],estNull[2])))
obsTest <- 2*(ell1-ellNull)
nboot = 100
ParallelTest(nboot,TN2016,obsTest,minRank)
}
\references{
{
\insertRef{bee22}{LNPar}
}
}
\keyword{likelihood;}
\keyword{log-likelihood}
\keyword{mixture;}
\keyword{profile}
\keyword{ratio}
\keyword{test.}
