% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPtest.R
\name{LPtest}
\alias{LPtest}
\title{Testing for a Pareto tail}
\usage{
LPtest(n, muNull, sigmaNull, obsTest, nboot, minRank)
}
\arguments{
\item{n}{sample size}

\item{muNull}{lognormal expected value under the null hypothesis.}

\item{sigmaNull}{lognormal standard deviation under the null hypothesis.}

\item{obsTest}{value of the test statistics computed with the data under analysis.}

\item{nboot}{number of bootstrap replications.}

\item{minRank}{minimum possible rank of the threshold.}
}
\value{
A list with the following elements:

lr: simulated values of the llr test under the null hypothesis.

pval: p-value of the test.
}
\description{
This function computes the bootstrap test for the null hypothesis of
a pure lognormal distribution versus the alternative of a lognormal-Pareto
mixture.
}
\examples{
resTest <- LPtest(100,0,1,0.67,100,90)
}
\keyword{likelihood;}
\keyword{log-likelihood}
\keyword{mixture;}
\keyword{profile}
\keyword{ratio}
\keyword{test.}
